+++ package.json
@@ -16,7 +16,7 @@
     "semantic-release": "semantic-release"
   },
   "dependencies": {
-    "next": "12.3.0",
+    "next": "^13.0.3",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "sass": "^1.54.9"
