+++ public/static/pollarized.png
