+++ styles/Home.module.scss
@@ -1,8 +1,6 @@
-.container {
-  padding: 0 2rem;
-  display: flex;
-  flex-direction: column;
-  height: 100%;
+.title a {
+  color: #0070f3;
+  text-decoration: none;
 }
 
 .main {
@@ -12,29 +10,6 @@
   justify-content: center;
   align-items: center;
   flex-direction: column;
-}
-
-.footer {
-  display: flex;
-  flex: 1;
-  padding: 1.2rem 0;
-  border-top: 3px solid #eaeaea;
-  align-items: center;
-  justify-content: flex-end;
-  position: absolute;
-  width: 100%;
-}
-
-.footer a {
-  display: flex;
-  justify-content: center;
-  align-items: bottom;
-  flex-grow: 1;
-}
-
-.title a {
-  color: #0070f3;
-  text-decoration: none;
 }
 
 .title a:hover,
@@ -116,17 +91,6 @@
   line-height: 1.5;
 }
 
-.vlogo {
-  height: 1em;
-  margin-left: 0.6rem;
-}
-
-.logo {
-  margin-top: -0.8rem;
-  height: 1em;
-  margin-left: 0.6rem;
-}
-
 @media (max-width: 600px) {
   .grid {
     width: 100%;
@@ -135,8 +99,7 @@
 }
 
 @media (prefers-color-scheme: dark) {
-  .card,
-  .footer {
+  .card {
     border-color: #222;
   }
   .codetext {
@@ -146,7 +109,5 @@
     -webkit-text-fill-color: transparent;
     background-clip:text;
   }
-  .vlogo img {
-    filter: invert(1);
-  }
+
 }
